requirements.txt
langgraph>=0.2.0,<1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0,<1.0.0
pydantic>=2.0
python-dotenv>=1.0.0
pytest>=7.0
langfuse>=2.7.0,<3.0
